Complete overhaul of player classification system
The way the player's characeter values are inputted and stored has been completely redone.

It is now done by a proper class system which should allow an easier transistion to save systems in the future. It also makes calling the players sex, name or race much easier. And makes it easier to edit and add on.

The old intro and storage method is provided in a text file.

One bad thing about switching to this class system is that I don't know how to invalidate garbage inputs. So I've simply put a warning that typing invalid or garbage values will likely break the game. Hope to find a way to just completely prevent that from happening. Add another thing to the to do list.