Impletemented New World Space
Started implementing new world spaces.

Added function fronthousearea to act as the portal to the outside world from the starting house.

Implemented the framework to implement 2 future areas, the pool, and the cafe.

Implemented parts of the Sycamore Lake Park area. Currently implementing dialog for both baths.

At the park the players have the option to go on a walking path to the east or the west. The results from this are randomized. You will either get a scene involing the PC's rommates or a scene involing the backstory of the PC themselves.

If you get one scene you will be to expierence the other by going on the other path (Not currently implemented)