Completely re-worked intro. Dialogue and story refinement.
The intro has been completely changed. It gives a lot more info about the players 'disappearance' and avoids the cliché of waking up/the PC having memory loss. The player now starts off on vacation at a resort and then returns home. The game goes on as usual after that.