Fully implemented a working save and load system.
I have implemented a complete saving and loading system using pickle. The player can now save their progress and resume at a later time. All of their actions and character info will be saved.

This will allow for the game to be expanded even further since the game no longer has to be completed in one session.

With this completed most of what I would consider the 'core functions' of a game are complete. There's a working travel system and save system. There might be some QoL changes I make before moving on and implementing further areas. We'll see.

But for now there are no major hurdles holding me back from making this game the best it can be.