Major 'under the hood' changes. Complete re-do of most systems.
Save and travel system are completely redone and should, probably, maybe, be error free now. We'll see.
The save system and travel system are now all object-orientated. They are controlled by classes and objects in CharInfo and their own file.
There should be no more circular imports. The whole point of basically re-coding the whole game was to prevent this.
All checks/variables are object-orientated, contained in CharInfo but could easily be moved if needed. Again, this helps prevent circular imports.