Reworked some of the checks and saving points.
Fully implemented the baseball fair game. Made it so the player can do either the fair games first or the walkway first, whatever they want.

Made it so the player can only go to the fair games if they brought along Holly. It was just too much dialogue too implement other wise. The fair games probably wouldn't have served much purpose anyway, since they're mainly being used to build the relationship with Holly.

Overall the festival area is really close to being finished. Just need to implement the other fair game and it's pretty much ready. I'll have to test everything out once I get to that point, but other than that I don't foresee anything preventing me from releasing the first version. Which will be done once all the festival areas are put in.

I'll still be doing touches of work here and there on the festival. Probably dialogu changes mainly. But the next big project is going to be moving forward with the road trip. Move on that in a future blog.