Continued work on chapter 3
Made a new house/base file since the O.G one was getting way to cluttered and convoluted. This one will only be used for scenes in the house unlike the previous one, which should help prevent it from getting to big. Still working on implementing everything but so far it seems like it’ll be a lot easier to add new stuff.

Added the opening scene for the second half of chapter 3. Dialogue differs depending on ending points and your previous choices.