Added Version/Update Checker
Tales from the Road now has version checking implemented, and it was surprisingly easy to add. It was implemented using JSON and the requests module, the game now reads a file on my GitHub website to check to see if I’ve uploaded any new versions of the game. It looks a bit janky, but it works fine so I’m probably not going to sweat it too much. So starting with 0.1.5 you’ll notified at the loading screen whenever a new version of the game is available.