Filled in some placeholders I still had for the Valery scene
Tied up a lot of loose ends I had in the Valery scene. Game now knows what to do if you don't get Valery's true ending. That ending does not have nearly as much content as the true ending because I just don't have the time to add it.

Still wanna do a full sweep and maybe add in a few extra things.